requirements.txt
streamlit>=1.50.0
pandas
plotly
openpyxl
Office365-REST-Python-Client
